rename seed package to prg
