+++ consensus/hotstuff/committees/leader/consensus.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/onflow/flow-go/model/flow/filter"
 	"github.com/onflow/flow-go/state/protocol"
-	"github.com/onflow/flow-go/state/protocol/seed"
+	"github.com/onflow/flow-go/state/protocol/prg"
 )
 
 // SelectionForConsensus pre-computes and returns leaders for the consensus committee
@@ -26,7 +26,7 @@ func SelectionForConsensus(epoch protocol.Epoch) (*LeaderSelection, error) {
 		return nil, fmt.Errorf("could not get epoch seed: %w", err)
 	}
 	// create random number generator from the seed and customizer
-	rng, err := seed.PRGFromRandomSource(randomSeed, seed.ConsensusLeaderSelection)
+	rng, err := prg.FromRandomSource(randomSeed, prg.ConsensusLeaderSelection)
 	if err != nil {
 		return nil, fmt.Errorf("could not create rng: %w", err)
 	}
