+++ module/chunks/chunk_assigner.go
@@ -12,7 +12,7 @@ import (
 	"github.com/onflow/flow-go/module/mempool"
 	"github.com/onflow/flow-go/module/mempool/stdmap"
 	"github.com/onflow/flow-go/state/protocol"
-	"github.com/onflow/flow-go/state/protocol/seed"
+	"github.com/onflow/flow-go/state/protocol/prg"
 )
 
 // ChunkAssigner implements an instance of the Public Chunk Assignment
@@ -98,7 +98,7 @@ func (p *ChunkAssigner) rngByBlockID(stateSnapshot protocol.Snapshot) (random.Ra
 		return nil, fmt.Errorf("failed to retrieve source of randomness: %w", err)
 	}
 
-	rng, err := seed.PRGFromRandomSource(randomSource, seed.VerificationChunkAssignment)
+	rng, err := prg.FromRandomSource(randomSource, prg.VerificationChunkAssignment)
 	if err != nil {
 		return nil, fmt.Errorf("failed to instantiate random number generator: %w", err)
 	}
